fix: Update jspdf-autotable usage
The import for jspdf-autotable was changed from importing the function to importing the module. The usage of the autoTable function was updated to reflect this change, casting the doc to `any` to access the autoTable method. The `base` tag in `index.html` was also removed as it is no longer necessary with the current Vite configuration.